+++ TakemeHome0_0/src/main/java/pe/edu/upc/aww/takemehome0_0/dtos/ListPricesGreaterThan1000DTO.java
@@ -2,7 +2,7 @@ package pe.edu.upc.aww.takemehome0_0.dtos;
 
 public class ListPricesGreaterThan1000DTO {
     private String nameProduct;
-    private double priceProduct;
+    private int priceProduct;
 
     public String getNameProduct() {
         return nameProduct;
@@ -12,11 +12,11 @@ public class ListPricesGreaterThan1000DTO {
         this.nameProduct = nameProduct;
     }
 
-    public double getPriceProduct() {
+    public int getPriceProduct() {
         return priceProduct;
     }
 
-    public void setPriceProduct(double priceProduct) {
+    public void setPriceProduct(int priceProduct) {
         this.priceProduct = priceProduct;
     }
 }
